core: binaryfile: refactor _open_elf logic
This eliminates duplicate code for section retrieval.
It is now possible to know whether some of the sections
.got, .got.plt, .data and .rodata was not found in the file.

This fixes also a "reference-before-assignment" error when one
of the section listed above is missing.

NB. This does not add checks to see if one of those section is present,
the previous behaviour should be unchanged. (Except for the reference
before assignment part)

Signed-off-by: Seraphime Kirkovski <kirkseraph@gmail.com>